# routes/gamification.js
const express = require('express');
const router = express.Router();
const { authenticateJWT } = require('../middleware/auth');

// Get all badges
router.get('/badges', authenticateJWT, async (req, res) => {
  try {
    // TODO: Implement badges retrieval
    const badges = [
      { id: '1', name: 'First Task', description: 'Complete your first task', icon: '🎯' },
      { id: '2', name: 'Streak Master', description: 'Maintain a 7-day streak', icon: '🔥' },
      { id: '3', name: 'Team Player', description: 'Collaborate on 10 tasks', icon: '🤝' }
    ];
    
    res.json({
      message: 'Badges retrieved successfully',
      badges
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to fetch badges', error: err.message });
  }
});

// Get user badges
router.get('/badges/user/:id', authenticateJWT, async (req, res) => {
  try {
    const { id } = req.params;
    
    // TODO: Implement user badges retrieval
    const userBadges = [
      { id: '1', name: 'First Task', earnedAt: '2023-12-01', icon: '🎯' },
      { id: '2', name: 'Streak Master', earnedAt: '2023-12-15', icon: '🔥' }
    ];
    
    res.json({
      message: 'User badges retrieved successfully',
      userId: id,
      badges: userBadges
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to fetch user badges', error: err.message });
  }
});

// Award badge to user
router.post('/badges/award', authenticateJWT, async (req, res) => {
  try {
    const { userId, badgeId, reason } = req.body;
    
    // TODO: Implement badge awarding
    const awardedBadge = {
      userId,
      badgeId,
      awardedBy: req.user._id,
      reason,
      awardedAt: new Date().toISOString()
    };
    
    res.json({
      message: 'Badge awarded successfully',
      awardedBadge
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to award badge', error: err.message });
  }
});

// Get all achievements
router.get('/achievements', authenticateJWT, async (req, res) => {
  try {
    // TODO: Implement achievements retrieval
    const achievements = [
      { id: '1', name: 'Task Master', description: 'Complete 100 tasks', icon: '🏆' },
      { id: '2', name: 'Efficiency Expert', description: 'Maintain 90% efficiency for a month', icon: '⚡' },
      { id: '3', name: 'Collaboration Champion', description: 'Work with 20 different team members', icon: '🌟' }
    ];
    
    res.json({
      message: 'Achievements retrieved successfully',
      achievements
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to fetch achievements', error: err.message });
  }
});

// Unlock achievement
router.post('/achievements/unlock', authenticateJWT, async (req, res) => {
  try {
    const { userId, achievementId, unlockedAt } = req.body;
    
    // TODO: Implement achievement unlocking
    const unlockedAchievement = {
      userId,
      achievementId,
      unlockedAt: unlockedAt || new Date().toISOString(),
      points: 100
    };
    
    res.json({
      message: 'Achievement unlocked successfully',
      unlockedAchievement
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to unlock achievement', error: err.message });
  }
});

// Get leaderboard
router.get('/leaderboard', authenticateJWT, async (req, res) => {
  try {
    // TODO: Implement global leaderboard
    const leaderboard = [
      { rank: 1, userId: 'user1', name: 'John Doe', points: 2500, level: 25 },
      { rank: 2, userId: 'user2', name: 'Jane Smith', points: 2200, level: 22 },
      { rank: 3, userId: 'user3', name: 'Bob Johnson', points: 2000, level: 20 }
    ];
    
    res.json({
      message: 'Leaderboard retrieved successfully',
      leaderboard
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to fetch leaderboard', error: err.message });
  }
});

// Get team leaderboard
router.get('/leaderboard/team', authenticateJWT, async (req, res) => {
  try {
    const { teamId } = req.query;
    
    // TODO: Implement team leaderboard
    const teamLeaderboard = [
      { rank: 1, userId: 'user1', name: 'John Doe', points: 1200, tasksCompleted: 45 },
      { rank: 2, userId: 'user2', name: 'Jane Smith', points: 1100, tasksCompleted: 42 },
      { rank: 3, userId: 'user3', name: 'Bob Johnson', points: 1000, tasksCompleted: 38 }
    ];
    
    res.json({
      message: 'Team leaderboard retrieved successfully',
      teamId,
      leaderboard: teamLeaderboard
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to fetch team leaderboard', error: err.message });
  }
});

// Get department leaderboard
router.get('/leaderboard/department', authenticateJWT, async (req, res) => {
  try {
    const { departmentId } = req.query;
    
    // TODO: Implement department leaderboard
    const departmentLeaderboard = [
      { rank: 1, userId: 'user1', name: 'John Doe', points: 1800, department: 'Engineering' },
      { rank: 2, userId: 'user2', name: 'Jane Smith', points: 1700, department: 'Engineering' },
      { rank: 3, userId: 'user3', name: 'Bob Johnson', points: 1600, department: 'Design' }
    ];
    
    res.json({
      message: 'Department leaderboard retrieved successfully',
      departmentId,
      leaderboard: departmentLeaderboard
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to fetch department leaderboard', error: err.message });
  }
});

// Get weekly leaderboard
router.get('/leaderboard/weekly', authenticateJWT, async (req, res) => {
  try {
    const { week } = req.query;
    
    // TODO: Implement weekly leaderboard
    const weeklyLeaderboard = [
      { rank: 1, userId: 'user1', name: 'John Doe', points: 500, week: week || 'current' },
      { rank: 2, userId: 'user2', name: 'Jane Smith', points: 450, week: week || 'current' },
      { rank: 3, userId: 'user3', name: 'Bob Johnson', points: 400, week: week || 'current' }
    ];
    
    res.json({
      message: 'Weekly leaderboard retrieved successfully',
      week: week || 'current',
      leaderboard: weeklyLeaderboard
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to fetch weekly leaderboard', error: err.message });
  }
});

// Get monthly leaderboard
router.get('/leaderboard/monthly', authenticateJWT, async (req, res) => {
  try {
    const { month, year } = req.query;
    
    // TODO: Implement monthly leaderboard
    const monthlyLeaderboard = [
      { rank: 1, userId: 'user1', name: 'John Doe', points: 2000, month: month || 'current' },
      { rank: 2, userId: 'user2', name: 'Jane Smith', points: 1800, month: month || 'current' },
      { rank: 3, userId: 'user3', name: 'Bob Johnson', points: 1600, month: month || 'current' }
    ];
    
    res.json({
      message: 'Monthly leaderboard retrieved successfully',
      month: month || 'current',
      year: year || new Date().getFullYear(),
      leaderboard: monthlyLeaderboard
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to fetch monthly leaderboard', error: err.message });
  }
});

// Get user streaks
router.get('/streaks/user/:id', authenticateJWT, async (req, res) => {
  try {
    const { id } = req.params;
    
    // TODO: Implement user streaks retrieval
    const userStreaks = {
      currentStreak: 5,
      longestStreak: 12,
      totalStreaks: 8,
      lastActivity: '2023-12-20'
    };
    
    res.json({
      message: 'User streaks retrieved successfully',
      userId: id,
      streaks: userStreaks
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to fetch user streaks', error: err.message });
  }
});

// Add points to user
router.post('/points/add', authenticateJWT, async (req, res) => {
  try {
    const { userId, points, reason, source } = req.body;
    
    // TODO: Implement points addition
    const pointsTransaction = {
      userId,
      points,
      reason,
      source,
      addedBy: req.user._id,
      addedAt: new Date().toISOString(),
      newBalance: 2500 // TODO: Calculate actual balance
    };
    
    res.json({
      message: 'Points added successfully',
      pointsTransaction
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to add points', error: err.message });
  }
});

// Get points history
router.get('/points/history', authenticateJWT, async (req, res) => {
  try {
    const { userId } = req.query;
    
    // TODO: Implement points history retrieval
    const pointsHistory = [
      { id: '1', points: 100, reason: 'Task completed', source: 'task', date: '2023-12-20' },
      { id: '2', points: 50, reason: 'Streak maintained', source: 'streak', date: '2023-12-19' },
      { id: '3', points: 200, reason: 'Achievement unlocked', source: 'achievement', date: '2023-12-18' }
    ];
    
    res.json({
      message: 'Points history retrieved successfully',
      userId,
      history: pointsHistory
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to fetch points history', error: err.message });
  }
});

// Update user streaks
router.post('/streaks/update', authenticateJWT, async (req, res) => {
  try {
    const { userId, action, date } = req.body;
    
    // TODO: Implement streak update
    const streakUpdate = {
      userId,
      action,
      date: date || new Date().toISOString(),
      updatedAt: new Date().toISOString(),
      newStreak: 6 // TODO: Calculate actual streak
    };
    
    res.json({
      message: 'Streak updated successfully',
      streakUpdate
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to update streak', error: err.message });
  }
});

module.exports = router; 