# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Task Manager API",
    "version": "1.0.0",
    "description": "Comprehensive API documentation for the AI Task Manager SaaS platform with task management, AI features, team management, gamification, analytics, and more."
  },
  "servers": [
    { "url": "http://localhost:3000/api" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Badge": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "icon": { "type": "string" },
          "category": { "type": "string", "enum": ["achievement", "milestone", "special", "team", "department"] },
          "points": { "type": "number", "minimum": 0 },
          "rarity": { "type": "string", "enum": ["common", "uncommon", "rare", "epic", "legendary"] },
          "criteria": { "type": "string" },
          "isActive": { "type": "boolean", "default": true }
        },
        "required": ["name", "description", "category"]
      },
      "UserBadge": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "badgeId": { "type": "string" },
          "awardedAt": { "type": "string", "format": "date-time" },
          "awardedBy": { "type": "string" },
          "reason": { "type": "string" },
          "badge": { "$ref": "#/components/schemas/Badge" }
        },
        "required": ["userId", "badgeId", "awardedAt"]
      },
      "AwardedBadge": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "badgeId": { "type": "string" },
          "awardedAt": { "type": "string", "format": "date-time" },
          "awardedBy": { "type": "string" },
          "reason": { "type": "string" },
          "message": { "type": "string" }
        },
        "required": ["userId", "badgeId", "awardedAt"]
      },
      "Achievement": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "icon": { "type": "string" },
          "category": { "type": "string", "enum": ["productivity", "collaboration", "learning", "leadership", "innovation"] },
          "points": { "type": "number", "minimum": 0 },
          "criteria": { "type": "string" },
          "isActive": { "type": "boolean", "default": true }
        },
        "required": ["name", "description", "category"]
      },
      "UnlockedAchievement": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "achievementId": { "type": "string" },
          "unlockedAt": { "type": "string", "format": "date-time" },
          "achievement": { "$ref": "#/components/schemas/Achievement" }
        },
        "required": ["userId", "achievementId", "unlockedAt"]
      },
      "LeaderboardEntry": {
        "type": "object",
        "properties": {
          "rank": { "type": "integer", "minimum": 1 },
          "userId": { "type": "string" },
          "userName": { "type": "string" },
          "userEmail": { "type": "string" },
          "points": { "type": "number", "minimum": 0 },
          "badges": { "type": "integer", "minimum": 0 },
          "achievements": { "type": "integer", "minimum": 0 },
          "streaks": { "type": "integer", "minimum": 0 }
        },
        "required": ["rank", "userId", "userName", "points"]
      },
      "UserStreaks": {
        "type": "object",
        "properties": {
          "userId": { "type": "string" },
          "currentStreak": { "type": "integer", "minimum": 0 },
          "longestStreak": { "type": "integer", "minimum": 0 },
          "totalStreaks": { "type": "integer", "minimum": 0 },
          "lastActivity": { "type": "string", "format": "date-time" },
          "streakType": { "type": "string", "enum": ["daily", "weekly", "monthly"] },
          "nextMilestone": { "type": "integer" }
        },
        "required": ["userId", "currentStreak", "longestStreak"]
      },
      "PointsTransaction": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "points": { "type": "number" },
          "reason": { "type": "string" },
          "source": { "type": "string", "enum": ["task", "achievement", "streak", "bonus", "other"] },
          "transactionDate": { "type": "string", "format": "date-time" },
          "balance": { "type": "number" }
        },
        "required": ["userId", "points", "reason", "source"]
      },
      "PointsHistoryEntry": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "points": { "type": "number" },
          "reason": { "type": "string" },
          "source": { "type": "string", "enum": ["task", "achievement", "streak", "bonus", "other"] },
          "transactionDate": { "type": "string", "format": "date-time" },
          "balance": { "type": "number" }
        },
        "required": ["userId", "points", "reason", "source"]
      },
      "StreakUpdate": {
        "type": "object",
        "properties": {
          "userId": { "type": "string" },
          "action": { "type": "string", "enum": ["increment", "reset", "maintain"] },
          "previousStreak": { "type": "integer", "minimum": 0 },
          "newStreak": { "type": "integer", "minimum": 0 },
          "updatedAt": { "type": "string", "format": "date-time" },
          "message": { "type": "string" }
        },
        "required": ["userId", "action", "newStreak"]
      },
...
    }
  },
  "security": [
    { "bearerAuth": [] }
  ],
...
  "paths": {
...
    "/badges": {
      "get": {
        "tags": ["Gamification"],
        "summary": "Get all available badges",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { 
            "description": "Badges retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "badges": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Badge"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/badges/user/{id}": {
      "get": {
        "tags": ["Gamification"],
        "summary": "Get badges earned by a specific user",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { 
            "description": "User badges retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "userId": { "type": "string" },
                    "badges": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserBadge"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/badges/award": {
      "post": {
        "tags": ["Gamification"],
        "summary": "Award a badge to a user",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": { "type": "string" },
                  "badgeId": { "type": "string" },
                  "reason": { "type": "string" }
                },
                "required": ["userId", "badgeId"]
              }
            }
          }
        },
        "responses": {
          "200": { 
            "description": "Badge awarded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "awardedBadge": {
                      "$ref": "#/components/schemas/AwardedBadge"
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User or badge not found" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/achievements": {
      "get": {
        "tags": ["Gamification"],
        "summary": "Get all available achievements",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { 
            "description": "Achievements retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "achievements": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Achievement"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/achievements/unlock": {
      "post": {
        "tags": ["Gamification"],
        "summary": "Unlock an achievement for a user",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": { "type": "string" },
                  "achievementId": { "type": "string" },
                  "unlockedAt": { "type": "string", "format": "date-time" }
                },
                "required": ["userId", "achievementId"]
              }
            }
          }
        },
        "responses": {
          "200": { 
            "description": "Achievement unlocked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "unlockedAchievement": {
                      "$ref": "#/components/schemas/UnlockedAchievement"
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User or achievement not found" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/leaderboard": {
      "get": {
        "tags": ["Gamification"],
        "summary": "Get global leaderboard",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10, "minimum": 1, "maximum": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0, "minimum": 0 } }
        ],
        "responses": {
          "200": { 
            "description": "Leaderboard retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "leaderboard": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LeaderboardEntry"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/leaderboard/team": {
      "get": {
        "tags": ["Gamification"],
        "summary": "Get team leaderboard",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "teamId", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10, "minimum": 1, "maximum": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0, "minimum": 0 } }
        ],
        "responses": {
          "200": { 
            "description": "Team leaderboard retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "teamId": { "type": "string" },
                    "leaderboard": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LeaderboardEntry"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Team ID required" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Team not found" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/leaderboard/department": {
      "get": {
        "tags": ["Gamification"],
        "summary": "Get department leaderboard",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "departmentId", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10, "minimum": 1, "maximum": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0, "minimum": 0 } }
        ],
        "responses": {
          "200": { 
            "description": "Department leaderboard retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "departmentId": { "type": "string" },
                    "leaderboard": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LeaderboardEntry"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Department ID required" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Department not found" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/leaderboard/weekly": {
      "get": {
        "tags": ["Gamification"],
        "summary": "Get weekly leaderboard",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "week", "in": "query", "schema": { "type": "string", "description": "Week identifier (e.g., '2023-W52')" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10, "minimum": 1, "maximum": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0, "minimum": 0 } }
        ],
        "responses": {
          "200": { 
            "description": "Weekly leaderboard retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "week": { "type": "string" },
                    "leaderboard": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LeaderboardEntry"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/leaderboard/monthly": {
      "get": {
        "tags": ["Gamification"],
        "summary": "Get monthly leaderboard",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "month", "in": "query", "schema": { "type": "string", "description": "Month (1-12)" } },
          { "name": "year", "in": "query", "schema": { "type": "string", "description": "Year (e.g., 2023)" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10, "minimum": 1, "maximum": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0, "minimum": 0 } }
        ],
        "responses": {
          "200": { 
            "description": "Monthly leaderboard retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "month": { "type": "string" },
                    "year": { "type": "string" },
                    "leaderboard": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LeaderboardEntry"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/streaks/user/{id}": {
      "get": {
        "tags": ["Gamification"],
        "summary": "Get user streaks information",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { 
            "description": "User streaks retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "userId": { "type": "string" },
                    "streaks": {
                      "$ref": "#/components/schemas/UserStreaks"
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/points/add": {
      "post": {
        "tags": ["Gamification"],
        "summary": "Add points to a user",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": { "type": "string" },
                  "points": { "type": "number", "minimum": 1 },
                  "reason": { "type": "string" },
                  "source": { "type": "string", "enum": ["task", "achievement", "streak", "bonus", "other"] }
                },
                "required": ["userId", "points", "reason"]
              }
            }
          }
        },
        "responses": {
          "200": { 
            "description": "Points added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "pointsTransaction": {
                      "$ref": "#/components/schemas/PointsTransaction"
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/points/history": {
      "get": {
        "tags": ["Gamification"],
        "summary": "Get points history for a user",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "userId", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0, "minimum": 0 } }
        ],
        "responses": {
          "200": { 
            "description": "Points history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "userId": { "type": "string" },
                    "history": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PointsHistoryEntry"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "User ID required" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/streaks/update": {
      "post": {
        "tags": ["Gamification"],
        "summary": "Update user streaks",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": { "type": "string" },
                  "action": { "type": "string", "enum": ["increment", "reset", "maintain"] },
                  "date": { "type": "string", "format": "date-time" }
                },
                "required": ["userId", "action"]
              }
            }
          }
        },
        "responses": {
          "200": { 
            "description": "Streak updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "streakUpdate": {
                      "$ref": "#/components/schemas/StreakUpdate"
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" },
          "500": { "description": "Internal server error" }
        }
      }
    },
...
  }
}